py/pytesseractDemo/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.2.1
pytesseract==0.3.13
